src/app/content-editable/content-editable.component.spec.ts
import { SimpleChange } from '@angular/core';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { ReactiveFormsModule } from '@angular/forms';
import { ContentEditableComponent } from './content-editable.component';

const TEST_STRING = 'test string.';

describe('ContentEditableComponent', () => {
  let component: ContentEditableComponent;
  let fixture: ComponentFixture<ContentEditableComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ ContentEditableComponent ],
      imports: [ ReactiveFormsModule ],
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ContentEditableComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should render empty component', async(() => {
    const compiled = fixture.debugElement.nativeElement;
    fixture.detectChanges();
    expect(compiled.textContent).toEqual(' ');
  }));

  it('should render content when editing is disabled', async(() => {
    component.model = TEST_STRING;
    component.enabled = false;
    component.ngOnChanges({
      'model': new SimpleChange(undefined, TEST_STRING, true),
      'enabled': new SimpleChange(undefined, false, true)
    });
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.textContent).toEqual(TEST_STRING);
  }));
});
